# api/openapi.yaml
openapi: 3.1.0
info:
  title: cyp-agent HTTP API
  version: 1.7.0
  description: |
    React 仪表盘使用的 REST 与 SSE 契约。

    真实 CEX 下单和 `live` 模式均为 hard-disabled。执行仅允许本地 Paper，或在
    `mode=paper`、`execution_venue=okx`、Demo 标识和完整 Demo 凭据同时满足时使用
    OKX 模拟环境；OKX Demo 当前只开放 USDT 线性永续合约。

    交易金额、价格、数量、PnL、费率和风险阈值在响应 JSON 中使用十进制字符串，
    避免经过二进制浮点。回测研究指标和运行计数仍使用 JSON number。集合与嵌套
    对象不返回 null；可选对象在不存在时省略。

    所有写请求必须使用 `application/json`。服务监听非回环地址时必须配置
    `CYP_API_TOKEN`，并通过 Bearer token（或 `X-CYP-API-Token`）认证。
servers:
  - url: /
    description: 与仪表盘同源
...
security: []
paths:
...
  /api/positions/close:
    post:
      tags: [Trading]
      security:
        - WriteToken: []
      operationId: closePosition
      summary: 按当前执行场所价格平仓
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosePositionRequest"
      responses:
        "200":
          description: 平仓执行结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    WriteToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: "`CYP_API_TOKEN`。回环监听且未配置 token 时，服务允许本机写请求。"
  responses:
    ErrorResponse:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          type: string
        code:
          type: string
    DecimalString:
      type: string
      description: 有限十进制；允许普通或指数记法，输出不使用 NaN/Infinity。
      pattern: "^[+-]?(?:[0-9]+(?:\\.[0-9]*)?|\\.[0-9]+)(?:[eE][+-]?[0-9]+)?$"
      examples: ["1000.00"]
...
    NullableDecimalString:
      oneOf:
        - $ref: "#/components/schemas/DecimalString"
        - type: "null"
...
    ProtectiveOrder:
      type: object
      additionalProperties: false
      required: [kind, order_id, trigger_price, reduce_only]
      properties:
        kind:
          type: string
          enum: [stop_loss, take_profit]
        order_id:
          type: string
        trigger_price: { $ref: "#/components/schemas/DecimalString" }
        reduce_only:
          type: boolean
    ExecutionResult:
      type: object
      additionalProperties: false
      required: [client_id, status, filled_base, fee_quote, protective_orders]
      properties:
        client_id:
          type: string
        order_id:
          type: [string, "null"]
        status:
          type: string
          enum:
            - new
            - preflight
            - submitting
            - acknowledged
            - partially_filled
            - filled
            - protective_placed
            - closed
            - canceled
            - rejected
            - failed
            - unknown
            - protective_failed
            - flattening
        filled_base: { $ref: "#/components/schemas/DecimalString" }
        avg_price: { $ref: "#/components/schemas/NullableDecimalString" }
        fee_quote: { $ref: "#/components/schemas/DecimalString" }
        slippage_bps: { $ref: "#/components/schemas/NullableDecimalString" }
        protective_orders:
          type: array
          items:
            $ref: "#/components/schemas/ProtectiveOrder"
        error:
          type: [string, "null"]
        chain:
          type: [string, "null"]
        tx_hash:
          type: [string, "null"]
        gas_used: { $ref: "#/components/schemas/NullableDecimalString" }
...
    ClosePositionRequest:
      type: object
      additionalProperties: false
      required: [symbol, instrument]
      properties:
        symbol:
          type: string
        instrument:
          type: string
          enum: [spot, perp]
...